wasm: move everything into common/ to match other extensions.
