+++ source/extensions/common/wasm/BUILD
@@ -17,6 +17,22 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
+    name = "config",
+    srcs = ["config.cc"],
+    hdrs = ["config.h"],
+    deps = [
+        "//include/envoy/registry",
+        "//include/envoy/server:wasm_config_interface",
+        "//source/common/common:assert_lib",
+        "//source/common/common:empty_string",
+        "//source/common/config:datasource_lib",
+        "//source/common/protobuf:utility_lib",
+        "//source/extensions/common/wasm:wasm_lib",
+        "//source/extensions/grpc_credentials:well_known_names",
+    ],
+)
+
+envoy_cc_library(
     name = "wasm_hdr",
     hdrs = ["wasm.h"],
     deps = [
