+++ source/extensions/common/wasm/wavm/BUILD
@@ -34,6 +34,6 @@ envoy_cc_library(
         "//source/common/protobuf",
         "//source/extensions/common/wasm:wasm_hdr",
         "//source/extensions/common/wasm:well_known_names",
-        "@envoy_api//envoy/config/wasm/v2:wasm_cc",
+        "@envoy_api//envoy/config/common/wasm/v2:wasm_cc",
     ],
 )
