+++ source/extensions/extensions_build_config.bzl
@@ -15,11 +15,6 @@ EXTENSIONS = {
     "envoy.grpc_credentials.file_based_metadata":      "//source/extensions/grpc_credentials/file_based_metadata:config",
 
     #
-    # WASM
-    #
-    "envoy.wasm":                                      "//source/extensions/wasm:config",
-
-    #
     # Health checkers
     #
