+++ test/extensions/common/wasm/BUILD
@@ -2,50 +2,45 @@ licenses(["notice"])  # Apache 2
 
 load(
     "//bazel:envoy_build_system.bzl",
+    "envoy_cc_test",
     "envoy_package",
-)
-load(
-    "//test/extensions:extensions_build_system.bzl",
-    "envoy_extension_cc_test",
 )
 
 envoy_package()
 
-envoy_extension_cc_test(
+envoy_cc_test(
     name = "wasm_test",
     srcs = ["wasm_test.cc"],
     data = [
-        "//test/extensions/wasm/test_data:modules",
+        "//test/extensions/common/wasm/test_data:modules",
     ],
-    extension_name = "envoy.wasm",
     external_deps = ["abseil_optional"],
     deps = [
         "//source/common/event:dispatcher_lib",
         "//source/common/stats:isolated_store_lib",
         "//source/common/stats:stats_lib",
+        "//source/extensions/common/wasm:config",
         "//source/extensions/common/wasm:wasm_lib",
-        "//source/extensions/wasm:config",
         "//test/test_common:environment_lib",
         "//test/test_common:simulated_time_system_lib",
     ],
 )
 
-envoy_extension_cc_test(
+envoy_cc_test(
     name = "config_test",
     srcs = ["config_test.cc"],
     data = [
-        "//test/extensions/wasm/test_data:modules",
+        "//test/extensions/common/wasm/test_data:modules",
     ],
-    extension_name = "envoy.wasm",
     deps = [
         "//include/envoy/registry",
         "//source/common/stats:isolated_store_lib",
+        "//source/extensions/common/wasm:config",
         "//source/extensions/common/wasm:wasm_lib",
-        "//source/extensions/wasm:config",
         "//source/server:wasm_config_lib",
         "//test/mocks/event:event_mocks",
         "//test/mocks/thread_local:thread_local_mocks",
         "//test/test_common:environment_lib",
-        "@envoy_api//envoy/config/wasm/v2:wasm_cc",
+        "@envoy_api//envoy/config/common/wasm/v2:wasm_cc",
     ],
 )
