+++ test/extensions/common/wasm/test_data/bad_signature.cc
