+++ test/extensions/common/wasm/test_data/logging.wasm
