+++ test/extensions/common/wasm/test_data/logging.wat
